--- Assets/KDJ/Scripts/SFX Manager.cs
@@ -1,125 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using System;
-using UnityEngine;
-
-public class SoundManager : MonoBehaviour
-{
-    [Serializable]
-    public struct SoundList
-    {
-        public string Name;
-        public AudioClip Clip;
-    }
-
-    [SerializeField] private List<SoundList> _bgmList = new List<SoundList>();
-    [SerializeField] private List<SoundList> _sfxList = new List<SoundList>();
-    [SerializeField] private AudioSource BGMPlayer;
-    [SerializeField] private AudioSource SFXPlayer;
-    [SerializeField] private Dictionary<string, AudioClip> BGMDic = new Dictionary<string, AudioClip>();
-    [SerializeField] private Dictionary<string, AudioClip> SFXDic = new Dictionary<string, AudioClip>();
-
-    public static SoundManager Instance;
-
-    private void Awake()
-    {
-        Init();
-    }
-
-    private void Init()
-    {
-        if (Instance == null)
-        {
-            Instance = this;
-            DontDestroyOnLoad(gameObject);
-        }
-        else
-        {
-            Destroy(gameObject);
-        }
-
-        ListInit();
-    }
-
-    /// <summary>
-    /// 실행 시 넣어놓은 BGM과 SFX를 Dictionary에 초기화합니다.
-    /// </summary>
-    private void ListInit()
-    {
-        foreach (var sound in _bgmList)
-        {
-            if (!BGMDic.ContainsKey(sound.Name))
-            {
-                BGMDic.Add(sound.Name, sound.Clip);
-            }
-        }
-
-        foreach (var sound in _sfxList)
-        {
-            if (!SFXDic.ContainsKey(sound.Name))
-            {
-                SFXDic.Add(sound.Name, sound.Clip);
-            }
-        }
-    }
-
-    /// <summary>
-    /// BGM을 재생합니다.
-    /// BGM이 재생 중일 경우, 새로 지정한 BGM으로 교체합니다.
-    /// </summary>
-    /// <param name="name">리스트에 있는 BGM의 이름</param>
-    public void PlayBGM(string name)
-    {
-        if (BGMDic.ContainsKey(name))
-        {
-            BGMPlayer.clip = BGMDic[name];
-            BGMPlayer.Play();
-        }
-    }
-
-    /// <summary>
-    /// SFX를 재생합니다.
-    /// </summary>
-    /// <param name="name">리스트에 있는 SFX의 이름</param>
-    public void PlaySFX(string name)
-    {
-        if (SFXDic.ContainsKey(name))
-        {
-            SFXPlayer.PlayOneShot(SFXDic[name]);
-        }
-    }
-
-    /// <summary>
-    /// SFX를 정지합니다.
-    /// </summary>
-    public void StopSFX()
-    {
-        SFXPlayer.Stop();
-    }
-
-    /// <summary>
-    /// BGM을 정지합니다.
-    /// </summary>
-    public void StopBGM()
-    {
-        BGMPlayer.Stop();
-    }
-
-    /// <summary>
-    /// BGM의 볼륨을 설정합니다.
-    /// </summary>
-    /// <param name="volume">BGM의 볼륨</param>
-    public void SetBGMVolume(float volume)
-    {
-        BGMPlayer.volume = volume;
-    }
-
-    /// <summary>
-    /// SFX의 볼륨을 설정합니다.
-    /// </summary>
-    /// <param name="volume">SFX의 볼륨</param>
-    public void SetSFXVolume(float volume)
-    {
-        SFXPlayer.volume = volume;
-    }
-}
