+++ Assets/KDJ/Scripts/SoundManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 2e64f7b157c33804ba38da7c5db44989
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
