+++ Assets/KDJ/Scripts/TestCode/TestSound.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: cc6f5d781260726448d125981c18ad08
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
